requirements_dev.txt
jinja2==2.11.3
django-jinja==2.6.0; python_version < '3.6'
django-jinja==2.7.0; python_version >= '3.6'
jinja2-simple-tags==0.3.1

pytest==6.1.2; python_version < '3.6'
pytest==6.2.2; python_version >= '3.6'
pytest-env==0.6.2
pytest-cov==2.11.1
pytest-django==3.10.0; python_version < '3.6'
pytest-django==4.1.0; python_version >= '3.6'
pytest-pythonpath==0.7.3
pytest-tldr==0.2.2
pytest-xdist==2.2.1
